Clean pcep/base-parser code
Fix maven checktyle errors:
 - Missing point at end of first comment for JavaDoc
 - Split line longer than 120 characters
 - Replace // fall-through by // fall through for OpenJDK-11 compliance
 - Rename Class XROSRLGSubobjectParser.java to XROSrlgSubobjectParser.java
 - Replace (short)X by Uint8.valueOf(X) for deprecated method

JIRA: BGPCEP-715

+++ pcep/base-parser/src/main/java/org/opendaylight/protocol/pcep/parser/subobject/XROIpv4PrefixSubobjectParser.java
@@ -28,9 +28,10 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.rsvp.rev
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.rsvp.rev150820.basic.explicit.route.subobjects.subobject.type.IpPrefixCase;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.rsvp.rev150820.basic.explicit.route.subobjects.subobject.type.IpPrefixCaseBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.rsvp.rev150820.basic.explicit.route.subobjects.subobject.type.ip.prefix._case.IpPrefixBuilder;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 /**
- * Parser for {@link IpPrefixCase}
+ * Parser for {@link IpPrefixCase}.
  */
 public class XROIpv4PrefixSubobjectParser implements XROSubobjectParser, XROSubobjectSerializer {
 
@@ -76,7 +77,7 @@ public class XROIpv4PrefixSubobjectParser implements XROSubobjectParser, XROSubo
             Preconditions.checkArgument(prefix.getIpv4Prefix() != null, "Ipv4Prefix is mandatory.");
             writeIpv4Prefix(prefix.getIpv4Prefix(), body);
             Preconditions.checkArgument(subobject.getAttribute() != null, "Attribute is mandatory.");
-            writeUnsignedByte((short) subobject.getAttribute().getIntValue(), body);
+            writeUnsignedByte(Uint8.valueOf(subobject.getAttribute().getIntValue()), body);
             XROSubobjectUtil.formatSubobject(TYPE, subobject.isMandatory(), body, buffer);
         }
     }
